requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
annotated-types==0.7.0
attrs==25.4.0
bcrypt==5.0.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
et_xmlfile==2.0.0
execnet==2.1.2
Faker==40.1.2
grpcio==1.76.0
grpcio-tools==1.76.0
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
jsonpath==0.82.2
loguru==0.7.3
lxml==6.0.2
more-itertools==10.8.0
openpyxl==3.1.5
packaging==25.0
paramiko==4.0.0
pluggy==1.6.0
premailer==3.10.0
protobuf==6.33.4
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
PyNaCl==1.6.2
pytest==9.0.2
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.19.1
schedule==1.2.2
setuptools==80.9.0
sshtunnel==0.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
xpinyin==0.7.7
yagmail==0.15.293
python-dotenv==1.0.1
